tugas 6: initial commit

+++ todolist/views.py
@@ -1,11 +1,13 @@
 from email import message
-from django.http import HttpResponseRedirect
-from django.shortcuts import render, redirect
+from django.core import serializers
+from django.http import HttpResponseRedirect, HttpResponse, JsonResponse
+from django.shortcuts import render, redirect, get_object_or_404
 from django.urls import reverse
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
+from django.views.decorators.csrf import csrf_exempt
 from todolist.models import Task
 from todolist.forms import InputForm
 import datetime
@@ -118,24 +120,57 @@ def create_task(request):
         )
 
 @login_required(login_url='/todolist/login/')
+@csrf_exempt
 def task_status(request, status_key):
-    task = Task.objects.get(
-        user = request.user,
-        pk = status_key
-        )
-    task.is_finished = not task.is_finished
-    task.save()
-    return redirect(
-        'todolist:show_todolist'
-        )
+    if request.method == 'POST':
+        task = get_object_or_404(Task, pk=status_key, user=request.user)
+        task.is_finished = not task.is_finished
+        task.save()
+        return JsonResponse({'error': False})
+    #task = Task.objects.get(
+        #user = request.user,
+        #pk = status_key
+        #)
+    #task.is_finished = not task.is_finished
+    #task.save()
+    #return redirect(
+    #    'todolist:show_todolist'
 
 @login_required(login_url='/todolist/login/')
+@csrf_exempt
 def delete_task(request, delete_key):
-    task = Task.objects.get(
-        user = request.user,
-        pk = delete_key
+    if request.method == 'POST':
+        task = get_object_or_404(Task, pk=delete_key, user=request.user)
+        task.is_finished = not task.is_finished
+        task.delete()
+        return JsonResponse({'error': False})
+    #task = Task.objects.get(
+    #    user = request.user,
+    #    pk = delete_key
+    #    )
+    #task.delete()
+    #return redirect(
+    #    'todolist:show_todolist'
+    #    )
+
+@login_required(login_url='/todolist/login/')
+def show_todolist_json(request):
+    data = Task.objects.filter( user=request.user ).order_by('id')
+    return HttpResponse(serializers.serialize('json', data),
+        content_type = 'application/json')
+
+@login_required(login_url='/todolist/login/')
+@csrf_exempt
+def add_task_ajax(request):
+    if request.method == 'POST':
+        title = request.POST['title']
+        description = request.POST['description']
+        task_item = Task.objects.create(
+            user=request.user, 
+            title=title, 
+            description=description,
         )
-    task.delete()
-    return redirect(
-        'todolist:show_todolist'
+        task_item.save()
+        return JsonResponse(
+                {"instance": "Task successfully added!"}, status=200
         )
